+++ config/splat.pspeu.dra.yaml
@@ -101,7 +101,6 @@ segments:
       - [0x23E60, c, dra_psp/23E60]
       - [0x23EF0, c, dra_psp/23EF0]
       - [0x23FE0, c, dra_psp/23FE0]
-      - [0x243F0, c, dra_psp/243F0]
       - [0x25670, c, dra_psp/25670]
       - [0x25720, c, dra_psp/25720]
       - [0x26170, c, dra_psp/26170]
@@ -200,7 +199,7 @@ segments:
       - [0xEE5C0, .rodata, dra_psp/1C030]
       - [0xEE600, .rodata, dra_psp/1DCE0]
       - [0xEE628, .rodata, dra_psp/1DDC0]
-      - [0xEE680, .rodata, dra_psp/243F0]
+      - [0xEE680, .rodata, dra_psp/23FE0]
       - [0xEE698, .rodata, dra_psp/25720]
       - [0xEE6C8, .rodata, dra_psp/262A0]
       - [0xEE708, .rodata, dra_psp/29760]
